Fix compiler warnings

+++ src/Filters/lanczos3filter.cpp
@@ -21,7 +21,7 @@
 #include <cmath>
 
 Lanczos3Filter::Lanczos3Filter() :
-    PI_FLOAT(std::atan(1.0) * 4)
+    PI_FLOAT(std::atanf(1.0) * 4)
 {
 }
 
@@ -36,8 +36,8 @@ float Lanczos3Filter::value(float value)
     if (value < 3.0f)
     {
         value *= PI_FLOAT;
-        double value_divided_by_3 = value / 3.0f;
-        return (std::sin(value) / value) * (std::sin(value_divided_by_3) / value_divided_by_3);
+        float value_divided_by_3 = value / 3.0f;
+        return (std::sinf(value) / value) * (std::sinf(value_divided_by_3) / value_divided_by_3);
     }
     return 0.0f;
 }
